refactor: moving into the deprecated.go
